EquidistantOnes.go
Check if all the 1’s in a binary string are equidistant or not

Given a binary string str, the task is to check if all the 1’s in the string are equidistant or not. The term equidistant means that the distance between every two adjacent 1’s is same. Note that the string contains at least two 1’s.